+++ app/src/main/java/com/example/rickandmortyapp/feature/characterscreen/ui/CharacterFragment.kt
@@ -6,26 +6,18 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
-import androidx.paging.compose.collectAsLazyPagingItems
 import com.example.rickandmortyapp.R
 import com.example.rickandmortyapp.core.ui.viewmodel.ViewModelFactory
-import com.example.rickandmortyapp.data.models.ListCharacterModel
 import com.example.rickandmortyapp.application.App
 import com.example.rickandmortyapp.domain.models.ListCharacter
 import com.example.rickandmortyapp.feature.episodescreen.ui.EpisodeFragment
-import com.example.rickandmortyapp.feature.characterscreen.ui.recycler.CharacterAdapter
-import com.example.rickandmortyapp.feature.characterscreen.ui.recycler.PageLoaderScrollListener
 import javax.inject.Inject
 
 class CharacterFragment : Fragment() {
-//
-//    private var _binding: FragmentCharacterBinding? = null
-//    private val binding get() = _binding ?: throw NullPointerException("Binding is not initialized")
+
     private lateinit var composeView: ComposeView
 
     @Inject
@@ -33,10 +25,6 @@ class CharacterFragment : Fragment() {
     private val viewModel: CharacterViewModel by viewModels {
         viewModelFactory
     }
-
-    private lateinit var adapter: CharacterAdapter
-
-    private lateinit var paginScrollListener: PageLoaderScrollListener
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
@@ -49,23 +37,6 @@ class CharacterFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-//        _binding = FragmentCharacterBinding.inflate(inflater, container, false)
-
-//        adapter = CharacterAdapter(this::openSecondScreen)
-//        binding.recycler.adapter = adapter
-
-        //toolbar
-//        binding.customToolbar.backButton.visibility = View.GONE
-//        binding.customToolbar.textTitle.text = getString(R.string.character)
-
-        //pagination
-//        paginScrollListener = PageLoaderScrollListener(
-//            binding.recycler.layoutManager as LinearLayoutManager,
-//            viewModel::loadData
-//        )
-//        binding.recycler.addOnScrollListener(paginScrollListener)
-
-//        return binding.root
         return ComposeView(requireContext()).also {
             composeView = it
         }
@@ -84,7 +55,7 @@ class CharacterFragment : Fragment() {
 
     }
 
-    fun openSecondScreen(model: ListCharacter) {
+    private fun openSecondScreen(model: ListCharacter) {
         findNavController().navigate(R.id.action_generalFragment_to_secondFragment,
             EpisodeFragment.dataForScreen(model)
         )
