+++ app/src/main/java/com/example/rickandmortyapp/feature/characterscreen/ui/characterlist/CharacterList.kt
@@ -1,4 +1,4 @@
-package com.example.rickandmortyapp.feature.characterscreen.ui
+package com.example.rickandmortyapp.feature.characterscreen.ui.characterlist
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.lazy.grid.GridCells
@@ -13,8 +13,8 @@ import androidx.paging.compose.collectAsLazyPagingItems
 import com.example.rickandmortyapp.R
 import com.example.rickandmortyapp.core.ui.compose.MainTitle
 import com.example.rickandmortyapp.domain.models.ListCharacter
-import com.example.rickandmortyapp.feature.characterscreen.ui.SampleData.flowFakeData
-import com.example.rickandmortyapp.feature.characterscreen.ui.SampleData.listData
+import com.example.rickandmortyapp.core.SampleData.flowFakeData
+import com.example.rickandmortyapp.core.SampleData.listData
 import kotlinx.coroutines.flow.Flow
 
 @Composable
